Code cleanup Compatibility restored with new server
