+++ src/main/java/fr/charlotte/arssdk/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package fr.colin.arssdk;
+package fr.charlotte.arssdk;
 
 public class UserNotFoundException extends Exception {
 }
