+++ src/main/java/fr/charlotte/arssdk/objects/CheckVessel.java
@@ -1,0 +1,17 @@
+package fr.charlotte.arssdk.objects;
+
+
+public class CheckVessel {
+
+    private String vesselID;
+    private String coID;
+    private String template;
+
+    public CheckVessel(String vesselID, String coID, String template) {
+        this.vesselID = vesselID;
+        this.coID = coID;
+        this.template = template;
+    }
+
+
+}
