+++ src/main/java/fr/charlotte/arssdk/objects/CheckVesselName.java
@@ -1,0 +1,15 @@
+package fr.charlotte.arssdk.objects;
+
+public class CheckVesselName {
+
+    private String vesselID;
+    private String coID;
+    private String text;
+
+    public CheckVesselName(String vesselID, String coID, String text) {
+        this.vesselID = vesselID;
+        this.coID = coID;
+        this.text = text;
+    }
+
+}
