--- src/main/java/fr/colin/arssdk/objects/CheckVessel.java
@@ -1,17 +1,0 @@
-package fr.colin.arssdk.objects;
-
-
-public class CheckVessel {
-
-    private String vesselid;
-    private String coid;
-    private String template;
-
-    public CheckVessel(String vesselid, String coid, String template) {
-        this.vesselid = vesselid;
-        this.coid = coid;
-        this.template = template;
-    }
-
-
-}
