--- src/main/java/fr/colin/arssdk/objects/CheckVesselName.java
@@ -1,15 +1,0 @@
-package fr.colin.arssdk.objects;
-
-public class CheckVesselName {
-
-    private String vesselid;
-    private String coid;
-    private String text;
-
-    public CheckVesselName(String vesselid, String coid, String text) {
-        this.vesselid = vesselid;
-        this.coid = coid;
-        this.text = text;
-    }
-
-}
